feat: Move business logic from UserDefaults to UserStorage class
